Fixed Shaders and made Point a double

+++ src/co/frostbyte/jge/entities/Entity.java
@@ -1,7 +1,7 @@
 package co.frostbyte.jge.entities;
 
 import co.frostbyte.jge.GameManager;
-import co.frostbyte.jge.Point;
+import co.frostbyte.jge.math.Point;
 
 public class Entity {
     private Animation animation;
@@ -26,18 +26,18 @@ public class Entity {
     }
 
     public void draw(int[] pixels) {
-        int maxWidth = point.getX() + animation.getCurrentFrame().getWidth();
-        int maxHeight = point.getY() + animation.getCurrentFrame().getHeight();
+        int maxWidth = (int) point.getX() + animation.getCurrentFrame().getWidth();
+        int maxHeight = (int) point.getY() + animation.getCurrentFrame().getHeight();
 
-        for (int x = point.getX(); x < maxWidth; x++) {
+        for (int x = (int) point.getX(); x < maxWidth; x++) {
             if (x < 0 || x > GameManager.WIDTH - 1)
                 break;
 
-            for (int y = point.getY(); y < maxHeight; y++) {
+            for (int y = (int) point.getY(); y < maxHeight; y++) {
                 if (y < 0 || y > GameManager.HEIGHT - 1)
                     continue;
 
-                int pixel = animation.getCurrentFrame().getPixels()[(x - point.getX()) + (y - point.getY()) * animation.getCurrentFrame().getWidth()];
+                int pixel = animation.getCurrentFrame().getPixels()[(x - (int) point.getX()) + (y - (int) point.getY()) * animation.getCurrentFrame().getWidth()];
 
                 if (pixel == 0) {
                     continue;
